requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.0
python-dotenv==1.0.0
langchain==0.1.17
langchain-openai==0.0.8
plotly==5.18.0